+++ wdtk-examples/src/main/java/org/wikidata/wdtk/examples/FetchOnlineDataExample.java
@@ -28,10 +28,11 @@ import org.wikidata.wdtk.datamodel.interfaces.EntityDocument;
 import org.wikidata.wdtk.datamodel.interfaces.ItemDocument;
 import org.wikidata.wdtk.datamodel.interfaces.PropertyIdValue;
 import org.wikidata.wdtk.wikibaseapi.WikibaseDataFetcher;
+import org.wikidata.wdtk.wikibaseapi.apierrors.MediaWikiApiErrorException;
 
 public class FetchOnlineDataExample {
 
-	public static void main(String[] args) {
+	public static void main(String[] args) throws MediaWikiApiErrorException {
 		ExampleHelpers.configureLogging();
 		printDocumentation();
 
@@ -39,6 +40,9 @@ public class FetchOnlineDataExample {
 
 		System.out.println("*** Fetching data for one entity:");
 		EntityDocument q42 = wbdf.getEntityDocument("Q42");
+		System.out
+				.println("The current revision of the data for entity Q42 is "
+						+ q42.getRevisionId());
 		if (q42 instanceof ItemDocument) {
 			System.out.println("The English name for entity Q42 is "
 					+ ((ItemDocument) q42).getLabels().get("en").getText());